+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - 添加实体字段。
   - com.dwarfeng.statistics.stack.bean.entity.StatisticsSetting.description。
+  - com.dwarfeng.statistics.stack.bean.dto.TaskUpdateModalInfo.messageUpdateFlag。
 
 - SPI 目录结构优化。
   - 将桥接机制的 SPI 接口与抽象类提相关代码文件提升至 `sdk` 模块中。